requirements.txt
json5
fastapi
uvicorn
tortoise-orm
